update to long

+++ src/main/java/com/huawei/hwcloud/gaussdb/data/store/race/DataStoreRaceImpl.java
@@ -50,7 +50,7 @@ public class DataStoreRaceImpl implements DataStoreRace {
     public void writeDeltaPacket(DeltaPacket deltaPacket) {
         try {
             writeCounter.add(1);
-            short count = deltaPacket.getDeltaCount();
+            long count = deltaPacket.getDeltaCount();
             long v = deltaPacket.getVersion();
             List<DeltaPacket.DeltaItem> list = deltaPacket.getDeltaItem();
             Map<Long, DeltaPacket.DeltaItem> map = new HashMap<>();
